More report tests to a `test_reporting.py`
