+++ unittests/test_policies.py
@@ -4,32 +4,22 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 import contextlib
-import json
-import jsonschema
 import os
 import pytest
-import socket
-import sys
-import time
 
 import reframe as rfm
 import reframe.core.runtime as rt
 import reframe.frontend.dependencies as dependencies
 import reframe.frontend.executors as executors
 import reframe.frontend.executors.policies as policies
-import reframe.frontend.reporting as reporting
-import reframe.utility.jsonext as jsonext
 import reframe.utility.osext as osext
 import unittests.utility as test_util
 
-from lxml import etree
 from reframe.core.exceptions import (AbortTaskError,
                                      FailureLimitError,
                                      ForceExitError,
-                                     ReframeError,
                                      RunSessionTimeout,
                                      TaskDependencyError)
-from reframe.frontend.reporting import RunReport
 from unittests.resources.checks.hellocheck import HelloTest
 from unittests.resources.checks.frontend_checks import (
     BadSetupCheck,
@@ -41,25 +31,6 @@ from unittests.resources.checks.frontend_checks import (
     SelfKillCheck,
     SystemExitCheck
 )
-
-
-# NOTE: We could move this to utility
-class timer:
-    '''Context manager for timing'''
-
-    def __init__(self):
-        self._time_start = None
-        self._time_end = None
-
-    def __enter__(self):
-        self._time_start = time.time()
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self._time_end = time.time()
-
-    def timestamps(self):
-        return self._time_start, self._time_end
 
 
 def make_kbd_check(phase='wait'):
@@ -80,25 +51,6 @@ def make_sleep_check():
         return test
 
     return _do_make_check
-
-
-@pytest.fixture
-def make_cases(make_loader):
-    def _make_cases(checks=None, sort=False, *args, **kwargs):
-        if checks is None:
-            checks = make_loader(
-                ['unittests/resources/checks'], *args, **kwargs
-            ).load_all(force=True)
-
-        cases = executors.generate_testcases(checks)
-        if sort:
-            depgraph, _ = dependencies.build_deps(cases)
-            dependencies.validate_deps(depgraph)
-            cases = dependencies.toposort(depgraph)
-
-        return cases
-
-    return _make_cases
 
 
 @pytest.fixture(params=['pre_setup', 'post_setup',
@@ -163,50 +115,9 @@ def num_failures_stage(runner, stage):
     return len([t for t in stats.failed() if t.failed_stage == stage])
 
 
-def _validate_runreport(report):
-    schema_filename = 'reframe/schemas/runreport.json'
-    with open(schema_filename) as fp:
-        schema = json.loads(fp.read())
-
-    jsonschema.validate(json.loads(report), schema)
-
-
-def _validate_junit_report(report):
-    # Cloned from
-    # https://raw.githubusercontent.com/windyroad/JUnit-Schema/master/JUnit.xsd
-    schema_file = 'reframe/schemas/junit.xsd'
-    with open(schema_file, encoding='utf-8') as fp:
-        schema = etree.XMLSchema(etree.parse(fp))
-
-    schema.assert_(report)
-
-
-def _generate_runreport(run_stats, time_start=None, time_end=None):
-    report = RunReport()
-    report.update_session_info({
-        'cmdline': ' '.join(sys.argv),
-        'config_files': rt.runtime().site_config.sources,
-        'data_version': reporting.DATA_VERSION,
-        'hostname': socket.gethostname(),
-        'prefix_output': rt.runtime().output_prefix,
-        'prefix_stage': rt.runtime().stage_prefix,
-        'user': osext.osuser(),
-        'version': osext.reframe_version(),
-        'workdir': os.getcwd()
-    })
-    if time_start and time_end:
-        report.update_timestamps(time_start, time_end)
-
-    if run_stats:
-        report.update_run_stats(run_stats)
-
-    return report
-
-
 def test_runall(make_runner, make_cases, common_exec_ctx, tmp_path):
     runner = make_runner()
-    with timer() as tm:
-        runner.runall(make_cases())
+    runner.runall(make_cases())
 
     assert 9 == runner.stats.num_cases()
     assert_runall(runner)
@@ -215,39 +126,6 @@ def test_runall(make_runner, make_cases, common_exec_ctx, tmp_path):
     assert 1 == num_failures_stage(runner, 'sanity')
     assert 1 == num_failures_stage(runner, 'performance')
     assert 1 == num_failures_stage(runner, 'cleanup')
-
-    # We dump the report first, in order to get any object conversions right
-    report = _generate_runreport(runner.stats, *tm.timestamps())
-    report.save(tmp_path / 'report.json')
-
-    # We explicitly set `time_total` to `None` in the last test case, in order
-    # to test the proper handling of `None`.
-    report['runs'][0]['testcases'][-1]['time_total'] = None
-
-    # Validate the junit report
-    _validate_junit_report(report.generate_xml_report())
-
-    # Read and validate the report using the `reporting` module
-    reporting.restore_session(tmp_path / 'report.json')
-
-    # Try to load a non-existent report
-    with pytest.raises(ReframeError, match='failed to load report file'):
-        reporting.restore_session(tmp_path / 'does_not_exist.json')
-
-    # Generate an invalid JSON
-    with open(tmp_path / 'invalid.json', 'w') as fp:
-        jsonext.dump(report, fp)
-        fp.write('invalid')
-
-    with pytest.raises(ReframeError, match=r'is not a valid JSON file'):
-        reporting.restore_session(tmp_path / 'invalid.json')
-
-    # Generate a report that does not comply to the schema
-    del report['session_info']['data_version']
-    report.save(tmp_path / 'invalid-version.json')
-    with pytest.raises(ReframeError,
-                       match=r'failed to validate report'):
-        reporting.restore_session(tmp_path / 'invalid-version.json')
 
 
 def test_runall_skip_system_check(make_runner, make_cases, common_exec_ctx):
@@ -339,18 +217,6 @@ def test_runall_skip_tests(make_runner, make_cases,
     more_cases, stage = make_cases_for_skipping()
     cases = make_cases() + more_cases
     runner.runall(cases)
-
-    def assert_reported_skipped(num_skipped):
-        run_report = _generate_runreport(runner.stats)['runs']
-        assert run_report[0]['num_skipped'] == num_skipped
-
-        num_reported = 0
-        for tc in run_report[0]['testcases']:
-            if tc['result'] == 'skip':
-                num_reported += 1
-
-        assert num_reported == num_skipped
-
     assert_runall(runner)
     assert 11 == runner.stats.num_cases(0)
     assert 5 == runner.stats.num_cases(1)
@@ -359,11 +225,9 @@ def test_runall_skip_tests(make_runner, make_cases,
     if stage.endswith('cleanup'):
         assert 0 == len(runner.stats.skipped(0))
         assert 0 == len(runner.stats.skipped(1))
-        assert_reported_skipped(0)
     else:
         assert 2 == len(runner.stats.skipped(0))
         assert 0 == len(runner.stats.skipped(1))
-        assert_reported_skipped(2)
 
 
 # We explicitly ask for a system with a non-local scheduler here, to make sure
@@ -478,18 +342,6 @@ def test_duration_limit(make_runner, make_cases, common_exec_ctx):
             assert num_aborted == 1
 
 
-@pytest.fixture
-def dep_checks(make_loader):
-    return make_loader(
-        ['unittests/resources/checks_unlisted/deps_complex.py']
-    ).load_all()
-
-
-@pytest.fixture
-def dep_cases(dep_checks, make_cases):
-    return make_cases(dep_checks, sort=True)
-
-
 def assert_dependency_run(runner):
     assert_runall(runner)
     stats = runner.stats
@@ -511,15 +363,16 @@ def assert_dependency_run(runner):
             assert os.path.exists(os.path.join(check.outputdir, 'out.txt'))
 
 
-def test_dependencies(make_runner, dep_cases, common_exec_ctx):
+def test_dependencies(make_runner, cases_with_deps, common_exec_ctx):
     runner = make_runner()
-    runner.runall(dep_cases)
+    runner.runall(cases_with_deps)
     assert_dependency_run(runner)
 
 
-def test_dependencies_with_retries(make_runner, dep_cases, common_exec_ctx):
+def test_dependencies_with_retries(make_runner, cases_with_deps,
+                                   common_exec_ctx):
     runner = make_runner(max_retries=2)
-    runner.runall(dep_cases)
+    runner.runall(cases_with_deps)
     assert_dependency_run(runner)
 
 
@@ -854,61 +707,6 @@ def test_compile_fail_reschedule_busy_loop(make_async_runner, make_cases,
     assert num_checks == stats.num_cases()
     assert_runall(runner)
     assert num_checks == len(stats.failed())
-
-
-@pytest.fixture
-def report_file(make_runner, dep_cases, common_exec_ctx, tmp_path):
-    runner = make_runner()
-    runner.policy.keep_stage_files = True
-    with timer() as tm:
-        runner.runall(dep_cases)
-
-    filename = tmp_path / 'report.json'
-    report = _generate_runreport(runner.stats, *tm.timestamps())
-    report.save(filename)
-    return filename
-
-
-def test_restore_session(report_file, make_runner,
-                         dep_cases, common_exec_ctx, tmp_path):
-    # Select a single test to run and create the pruned graph
-    selected = [tc for tc in dep_cases if tc.check.name == 'T1']
-    testgraph = dependencies.prune_deps(
-        dependencies.build_deps(dep_cases)[0], selected, max_depth=1
-    )
-
-    # Restore the required test cases
-    report = reporting.restore_session(report_file)
-    testgraph, restored_cases = report.restore_dangling(testgraph)
-
-    assert {tc.check.name for tc in restored_cases} == {'T4', 'T5'}
-
-    # Run the selected test cases
-    runner = make_runner()
-    with timer() as tm:
-        runner.runall(selected, restored_cases)
-
-    new_report = _generate_runreport(runner.stats, *tm.timestamps())
-    assert new_report['runs'][0]['num_cases'] == 1
-    assert new_report['runs'][0]['testcases'][0]['name'] == 'T1'
-
-    # Generate an empty report and load it as primary with the original report
-    # as a fallback, in order to test if the dependencies are still resolved
-    # correctly
-    empty_report = _generate_runreport(None, *tm.timestamps())
-    empty_report_file = tmp_path / 'empty.json'
-    empty_report.save(empty_report_file)
-
-    report2 = reporting.restore_session(empty_report_file, report_file)
-    restored_cases = report2.restore_dangling(testgraph)[1]
-    assert {tc.check.name for tc in restored_cases} == {'T4', 'T5'}
-
-    # Remove the test case dump file and retry
-    os.remove(tmp_path / 'stage' / 'generic' / 'default' /
-              'builtin' / 'T4' / '.rfm_testcase.json')
-
-    with pytest.raises(ReframeError, match=r'could not restore testcase'):
-        report.restore_dangling(testgraph)
 
 
 def test_config_params(make_runner, make_exec_ctx):
